+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add new cop `RSpec/LeakyLocalVariable`. ([@lovro-bikic])
 - Bump RuboCop requirement to +1.81. ([@ydah])
+- Fix a false positive for `RSpec/LetSetup` when `let!` used in outer scope. ([@ydah])
 
 ## 3.7.0 (2025-09-01)